Enable configurable article count
